+++ js/main.js
@@ -1,1 +1,2 @@
-import './util.js';
+import './utils.js';
+import './date.js';
